trying a different structure
Trying a different strucure as per:

http://stackoverflow.com/questions/3842616/organizing-python-classes-in-modules-and-or-packages
